refactor(p2p): Extract push_existing_docs to db crate, add DocPusher trait (#392 Phase 1)
Extract shared push logic from FFI into db::push_docs so CLI/HTTP can
reuse it. Add DocPusher trait + DbDocPusher to type-erase the store
parameter. Enrich P2PAdapter with event_bus, peer_addresses, and
tracked_documents fields needed by Phases 2-7.
